[feat] Migrate user related APIs
