+++ src/main.py
@@ -325,73 +325,6 @@ async def ingest_books_navbar(
     )
 
 
-@app.get("/profile", response_class=HTMLResponse)
-async def user_profile(
-    request: Request, db=Depends(get_db), user_info: str = Cookie(None)
-):
-    username = username_email_resol(user_info)
-    query = s_db.q_user_bks(username=username)
-    response = p.process_user_bks(query)
-    unique_libs = p.get_unique_libs(response)
-    mdb = m_db.connect_mdb()
-    mdb = mdb["nlb"]
-    update_status = None
-    if m_db.q_status(db=mdb, username=username):
-        update_status = " "
-
-    # Query user profile info from database
-    user_info = s_db.q_user_info(db, username)
-    return templates.TemplateResponse(
-        "profile.html",
-        {
-            "request": request,
-            "username": username,
-            "email_address": user_info.get("email_address", None),
-            "preferred_lib": user_info.get("preferred_lib", None),
-            "pw_qn": user_info.get("pw_qn", None),
-            "pw_ans": user_info.get("pw_ans", None),
-            "all_unique_lib": unique_libs,
-            "status": update_status,
-        },
-    )
-
-
-@app.post("/update_user", response_class=HTMLResponse)
-async def update_user(
-    request: Request,
-    email_address: str = Form(None),
-    preferred_lib: str = Form(None),
-    pw_qn: str = Form(None),
-    pw_ans: str = Form(None),
-    password: str = Form(None),
-    db=Depends(get_db),
-    user_info: str = Cookie(None),
-):
-    username = username_email_resol(user_info)
-    # Update info
-    new_dict = {
-        "email_address": email_address,
-        "preferred_lib": preferred_lib,
-        "pw_qn": pw_qn,
-        "pw_ans": pw_ans,
-    }
-    if password:
-        hashed_password = get_hashed_password(password)
-        new_dict.update({"HashedPassword": hashed_password})
-
-    s_db.update_user_info(db, username, new_dict)
-    return RedirectResponse(f"/profile", status_code=status.HTTP_302_FOUND)
-
-
-@app.post("/delete_user", response_class=HTMLResponse)
-async def delete_user(
-    request: Request, db=Depends(get_db), user_info: str = Cookie(None)
-):
-    username = username_email_resol(user_info)
-    s_db.delete_user(db, username=username)
-    return RedirectResponse("/logout", status_code=status.HTTP_302_FOUND)
-
-
 # Main Page
 @app.get("/", response_class=HTMLResponse)
 async def root(request: Request):
