+++ templates/partials/nb_profile_btn.html
@@ -1,5 +1,5 @@
 <li>
-    <a class="text-white hover:text-white hover:bg-sky-500" hx-get="/profile" hx-target=".main_content"
+    <a class="text-white hover:text-white hover:bg-sky-500" hx-get="/user/" hx-target=".main_content"
         hx-indicator="#head-spinner"><i class="fas fa-user"></i>Profile
     </a>
 </li>
